+++ README.md
@@ -32,8 +32,9 @@ ECONOVATION HEROES 2기 2기 Java Racing Game Practice
  - util 목적으로 사용하실 클래스 객체 생성 비용 줄이기
  - 의미없는 메모리 낭비 없애기
  - 가독성에 유리하게 선언과 할당 같은 곳에서 해주기
- ######해야할 일
- - 테스트 코드 작성
+ - 테스트 코드
+ ###### 해결할 문제
+
  테스트 코드를 작성하다보니 public으로만 테스트가 안된다는 점에서 코드가 잘 짜여지지 않다고 느꼈다.
  테스트 코드를 먼저 작성하고 그에 맞추어 코드를 작성해야할까?
