+++ test/RacingGameTest.java
@@ -7,41 +7,39 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 
-import static org.hamcrest.CoreMatchers.*;
+import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
 public class RacingGameTest {
-    List<Car> cars;
+        List<Car> cars;
+        RacingGame racingGame;
     @Before
     public void setUp() throws Exception {
         cars = new ArrayList<>();
         cars.add(new Car("애플"));
         cars.add(new Car("삼성"));
         cars.add(new Car("전남대"));
+        racingGame = new RacingGame();
     }
 
     @Test
-    public void inputCarName() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
-        String carName = "애플, 삼성, 전남대";
-        InputView inputView = new InputView();
-        try {
-            Method methodsplit = inputView.getClass().getDeclaredMethod("splitInputCarName", String.class);
-            methodsplit.setAccessible(true);
-            Method methodtrim = inputView.getClass().getDeclaredMethod("trimInputCarName", String[].class);
-            methodtrim.setAccessible(true);
-            String[] name = (String[]) methodsplit.invoke(inputView, carName);
-            String[] name1 = (String[]) methodtrim.invoke(inputView, (Object) name);
-            assertThat("애플", is(name1[0]));
-            assertThat("삼성", is(name1[1]));
-            assertThat("전남대", is(name1[2]));
-        } catch (NoSuchMethodException | IllegalAccessException | InvocationTargetException e) {
-            e.printStackTrace();
+    public void race() throws NoSuchMethodException, InvocationTargetException, IllegalAccessException {
+        Method goOrStay = racingGame.getClass().getDeclaredMethod("goOrStay", Car.class);
+        goOrStay.setAccessible(true);
+        for(Car car: cars) {
+            goOrStay.invoke(racingGame, car);
         }
     }
 
     @Test
-    public void inputTrial() {
-
+    public void getWinner() {
+        cars.get(0).goForward();
+        cars.get(1).goForward();
+        List<String> expectedWinnerNames = new ArrayList<>();
+        expectedWinnerNames.add("애플");
+        expectedWinnerNames.add("삼성");
+        List<String> actualWinnerNames = racingGame.getWinner(cars);
+        assertThat(actualWinnerNames,is(expectedWinnerNames));
     }
 
     @After
